--- a/packages/pocket-ic/test_canister/canister.did
+++ b/packages/pocket-ic/test_canister/canister.did
@@ -112,4 +112,5 @@ service : {
   blob_len : (blob) -> (nat64);
   call_with_large_blob : (principal, nat64) -> (nat64);
   execute_many_instructions : (nat64) -> ();
+  run : () -> ();
 }