This PR tries to reproduce a PocketIC performance issue when running memory-heavy query calls, but fails to do:
```
---- poc stdout ----
elapsed: 1.131458637s
elapsed: 1.135432527s
elapsed: 1.187247328s
elapsed: 1.16852671s
elapsed: 1.170953856s
elapsed: 1.156630414s
elapsed: 1.154909826s
elapsed: 1.129882688s
elapsed: 1.102209686s
elapsed: 1.144976877s
```
The benchmark durations stay constant.